Split runners into vm based runner and local evaluation
