+++ package.json
@@ -103,6 +103,7 @@
     "parse-diff": "^0.4.0",
     "parse-link-header": "^1.0.1",
     "pinpoint": "^1.1.0",
+    "require-from-string": "^2.0.1",
     "rfc6902": "^1.3.0",
     "vm2": "patriksimek/vm2#custom_files",
     "voca": "^1.2.0"
