+++ source/ambient.d.ts
@@ -14,3 +14,24 @@ declare module "parse-link-header"
 declare module "pinpoint"
 
 declare module "*/package.json"
+
+declare module "require-from-string"
+
+// declare module "require-from-string" {
+//   export interface RequireOptions {
+//     /** List of paths, that will be appended to module paths. Useful, when you want
+//      * to be able require modules from these paths. */
+//     appendPaths: string[]
+//     /**
+//      * Same as appendPath, but paths will be prepended.
+//      */
+//     prependPaths: string[]
+//   }
+//   /**
+//    * Load module from string in Node.
+//    * @param code Module code
+//    * @param filename Optional filename
+//    * @param opts
+//    */
+//   export default function(code: string, filename?: string, opts?: Partial<RequireOptions>): any
+// }
