+++ source/runner/_tests/_danger_runner.test.ts
@@ -1,5 +1,5 @@
 import { contextForDanger } from "../Dangerfile"
-import { createDangerfileRuntimeEnvironment, runDangerfileEnvironment, cleanDangerfile } from "../DangerfileRunner"
+import { createDangerfileRuntimeEnvironment, runDangerfileEnvironment } from "../runners/inline"
 
 import { FakeCI } from "../../ci_source/providers/Fake"
 import { FakePlatform } from "../../platforms/FakePlatform"
@@ -85,7 +85,7 @@ describe("with fixtures", () => {
     })
   })
 
-  it("handles multiple scheduled statements and all message types", async () => {
+  it.only("handles multiple scheduled statements and all message types", async () => {
     const context = await setupDangerfileContext()
     const runtime = await createDangerfileRuntimeEnvironment(context)
     const results = await runDangerfileEnvironment(
@@ -182,41 +182,5 @@ describe("with fixtures", () => {
 
     expect(results.fails[0].message).toContain("Danger failed to run")
     expect(results.markdowns[0]).toContain("Error: failure")
-  })
-})
-
-describe("cleaning Dangerfiles", () => {
-  it("also handles typescript style imports", () => {
-    const before = `
-import { danger, warn, fail, message } from 'danger'
-import { danger, warn, fail, message } from "danger"
-import { danger, warn, fail, message } from "danger";
-import danger from "danger"
-import danger from 'danger'
-import danger from 'danger';
-`
-    const after = `
-// Removed import
-// Removed import
-// Removed import
-// Removed import
-// Removed import
-// Removed import
-`
-    expect(cleanDangerfile(before)).toEqual(after)
-  })
-
-  it("also handles require style imports", () => {
-    const before = `
-const { danger, warn, fail, message } = require('danger')
-var { danger, warn, fail, message } = require("danger")
-let { danger, warn, fail, message } = require('danger');
-`
-    const after = `
-// Removed require
-// Removed require
-// Removed require
-`
-    expect(cleanDangerfile(before)).toEqual(after)
   })
 })
