+++ source/runner/_tests/fixtures/__DangerfileMultiScheduled.js
@@ -1,27 +1,37 @@
 /*eslint-disable */
 
-schedule(new Promise((res) => {
-  setTimeout(() => {
-    fail('Asynchronous Failure');
-    res();
-  }, 100);
-}));
+debugger
 
-schedule(new Promise((res) => {
-  warn('Asynchronous Warning');
-  res();
-}));
+schedule(
+  new Promise(res => {
+    setTimeout(() => {
+      fail("Asynchronous Failure")
+      res()
+    }, 100)
+  })
+)
 
-schedule(new Promise((res) => {
-  setTimeout(() => {
-    message('Asynchronous Message');
-    res();
-  }, 10);
-}));
+schedule(
+  new Promise(res => {
+    warn("Asynchronous Warning")
+    res()
+  })
+)
 
-schedule(new Promise((res) => {
-  setTimeout(() => {
-    markdown('Asynchronous Markdown');
-    res();
-  }, 50);
-}));
+schedule(
+  new Promise(res => {
+    setTimeout(() => {
+      message("Asynchronous Message")
+      res()
+    }, 10)
+  })
+)
+
+schedule(
+  new Promise(res => {
+    setTimeout(() => {
+      markdown("Asynchronous Markdown")
+      res()
+    }, 50)
+  })
+)
