+++ source/runner/runners/_tests/_transpiler.test.ts
@@ -1,0 +1,3 @@
+it("doesn't just fail yet", () => {
+  expect(1).toBe(1)
+})
