+++ source/runner/runners/utils/cleanDangerfile.ts
@@ -1,0 +1,12 @@
+// https://regex101.com/r/dUq4yB/1
+const requirePattern = /^.* require\(('|")danger('|")\);?$/gm
+//  https://regex101.com/r/dUq4yB/2
+const es6Pattern = /^.* from ('|")danger('|");?$/gm
+
+/**
+ * Updates a Dangerfile to remove the import for Danger
+ * @param {string} contents the file path for the dangerfile
+ * @returns {string} the revised Dangerfile
+ */
+export default (contents: string): string =>
+  contents.replace(es6Pattern, "// Removed import").replace(requirePattern, "// Removed require")
